JobRecord: remove cast to time.mktime()
Problem: The argument to time.mktime() is
the struct_time tuple which expresses the local
time. t_inactive and t_run are already expressed
as the seconds since the epoch, so there isn't
a need for this conversion.

Remove the conversion to time.mktime() for
the "elapsed" and "queued" properties
in a JobRecord object.