create-db: add job usage factor table
Add job_usage_factor_table, which will store
previous job usage factors per user/bank combination,
as well as the last t_inactive value found
for a user's jobs, so that when the next query for
job records is run, it will look for job records
that have completed after this timestamp.

Add t_half_life_period_table, which stores the
timestamp of the end of the current half-life period.
The timestamp is initially created at database creation
time, and gets updated with a new half-life period if
new jobs are found outside of the previous half-life
period.