fairness: add account_t class support
Introduce the base account_t class that will be used
to capture the state of either a bank account or user.
This class will initially be used by the upcoming
weighted_tree-based fairness implemenation.
But it should be also used by other future fairness
implemenations such that the same account base
schema is used for all of the fairness implementations.