fairness: library to provide weighted tree-based algorithm
Provide fairness based on the high-level
"hierarchical scheduling" concept explained in a paper
like https://people.eecs.berkeley.edu/~alig/papers/h-drf.pdf.
The algorithm is also similar to SLURM's fair tree
as explained in https://slurm.schedmd.com/fair_tree.html.

Two main classes are weighted_tree_node_t
and weighted_walk_t. weighted_tree_node_t
allows you to create a tree of bank accounts/users
(both called "accounts") and a weighted_walk_t object
will be created on the root of a tree and applies
recursive algorithms to calculate the fairness
of each and every user in a manner similar to
https://people.eecs.berkeley.edu/~alig/papers/h-drf.pdf
and https://slurm.schedmd.com/fair_tree.html.

Class weighted_tree_node_t derives from the
account_t class and adds support pertaining to
tree operations (e.g., a weighted_tree_node_t object
has the vector of weighted_tree_node_t objects as its
child nodes. We use share_ptr to make it easy to
manage memory.

An important method in this class is
calc_set_children_weight () which calculates
the weight for eacn and all of its children.
The weight of each child is the ratio of its
relative shares (with respect to its
siblings) to its relative usage.
The higher its given shares relative to
the combined shares of its siblings, and the lower the
given usage relative to the combined usage
of its siblings, its weight will get higher.
A higher weight account will have a higher priority
among the sibling accounts.

Class weighted_walk_t adds traversing operations
on the tree. Here, an important method is run () which
performs a depth first walk from the root of the tree such
a way that we visit a node with the highest
weight among the siblings of each level first.

The tie breaking algorithm is a bit tricky
and please refer to the comments in the methods:
run () and handle_internal ().