accounting: convert to a subpackage of the `flux` python package
Problem: the current setup installs `accounting` as its own top-level
package, which conflicts with other packages of the same name.  We would
prefer it to be installed as a subpackage of the `flux` package.  We
want this while retaining the ability to import it with `import
flux.accounting` both in-tree and when install.

Solution: mirror the python binding directory structure of flux-core,
specifically create a `src/bindings/python/flux` directory and make
`accounting` a subdirectory.  `PYTHONPATH` is then set to
`src/bindings/python` when in-tree, so that `import flux.accounting`
works as expected