flux-account: remove required db path CLI argument for create-db
Problem: flux-account now supports a compiled-in default of the install
location.  It isn't necessarily easy for the user to look up the
location of the `localstatedir` for their platform.  Additionally, the
`-p` argument is not respected for the `create-db` subcommand

Solution: remove the required database path CLI argument for
`create-db`. Instead use the compiled in default and respect the `-p`
user override if it is provided. Also, create any intermediate missing
directories, if required.