automake: add 'foreign' option
Problem: AUTHORS file is a pain to remember to maintain, is not
very compatible with our multi-repo "framework" structure, and is
for the most part, redundant with attributions in git commits.

It exists because automake insists that we have it to conform with
GNU project guidelines.

Set the 'foreign' automake option to relax enforcement of GNU project
guidelines.