job_usage_factor_table: remove ON DELETE CASCADE
We want to keep a user's job usage history values even
if they are removed from the flux-accounting DB, so
remove the ON DELETE CASCADE option from the creation
of job_usage_factor_table.

Add a test to confirm that even when a user is removed
from the flux-accounting DB, their entry in the
job_usage_factor_table is still present.

Fixes #84