github: edit linter step
The flake8 step currently reports
warnings and errors but does not
catch it as an error in GitHub
actions, so remove the --exit-zero
argument from the flake8 line and
add a number of ignored warnings
and error codes.

The black formatter step reports
formatting changes in the Python files
but also does not catch it as an error
in GitHub actions, so add a specific
version of the Black formatter and add
a --check argument to return a non-zero
exit code if there would be formatting
changes to files.