test: rename test DB files
Problem: Giving the mutliple test database files the same
name can cause issues with testing, for if the DB file doesn't
get deleted right away, the unit test will complain that the
DB already exists and the tests will fail.

Rename the sample DB files uniquely and according to the
type of unit tests they are working with.