t10000-print-hierarchy.t: change python3 command
Problem: The 'python3' command in some of the tests
could cause version mismatch issues if the Python version
of the sharness test was configured differently from the
Python version of the 'python3' command in a user's
environment.

Change the command to ${PYTHON}, which will call the version
of Python that the flux-accounting directory was configured
with.