calc_usage_factor tests: use unittest.mock ()
Two of the unit tests that interact with calc_usage_factor()
depend on time.time() to simulate a half-life period in order
to run their tests, which can cause a problem when comparing a
test value with its expected value.

Change t_end_half_life_period in the test database
to be a static time value. This will allow us to use
mock.MagicMock() to simulate time.time() values for
all of the tests in test_job_archive_interface.py.

Change test_14 and test_15 to no longer use time.time()
to create its time values.