remove print_hierarchy.py
As this repo matures, it feels like the C++ side of flux-accounting
deals with the the actual hierarchy of a flux-accounting DB, its
banks and users' weights and relationships with one another, and
its sorting as it relates to fair share. The Python side deals with
the administrative side of the flux-accounting database
(adding/removing users and banks, editing information, and the like).

Remove print_hierarchy.py from flux-accounting.