bionic: change pandas installation strategy
Problem: The bionic image installs the pandas
dependency under "/usr/lib/python3.7/site-packages/,"
which could cause confusion when configuring and building
flux-accounting with default options, shown by a
'ModuleNotFoundError' or a "missing dependency."

Change the the pandas installation location to the
default Python version on the bionic image, which is Python3.6.
This matches the other Docker image in flux-accounting, centos8.

Add autotools support to generate a symlink to
/usr/lib/python<version>/site-packages/ when flux-accounting
gets installed so that the fluxacct.accounting
namespace package gets recognized as a module when running
"flux account" commands.