calc_usage_factor: refactor/remove comments
The calc_usage_factor() and update_end_half_life_period()
functions has very long and some unneccesary comments
that make it hard to read.

Change some of the comments and remove some other
ones that I can just include in the documentation
for calc_usage_factor() without having to put it
directly in the source code.