calc_usage_factor: change decay & update behavior
Applying the decay factor to past usage bins and updating
usage values in the flux-accounting database in its current
state is incorrect in the following ways:

Problem (1): If new jobs are found in a new half-life period, then
apply_decay_factor() returns the sum of the past usage
bins INCLUDING the oldest bin. This is wrong because
in the case of a new half-life period, the oldest bin
should not be factored in when calculating the past usage
for a user.

Solution: Only return the sum of the usage bins NOT INCLUDING
the oldest bin.

---

Problem (2): If new jobs are found in the current half-life period,
only the FIRST usage bin (the one representing the current half-life
period) should be updated, NOT including past usage bins.

Solution: Add a new helper function, update_curr_usg_col(), which
updates the current usage bin (usage_factor_period_0) with the new
current usage factor. Place the calls to this helper function inside
the if-else branches where it is needed, which is the following
scenarios:

- when new jobs are found in the current half-life period
- when new jobs are found in the new half-life period

---

Problem (3): The historical job usage value gets updated even in the
scenario where no new jobs are found in the current half-life period.
This value should only be updated when a new half-life period is
detected OR when new jobs are found in the current half-life period.

Solution: Add a new helper function, update_hist_usg_col(), which updates
the job_usage column in the association_table with the historical
usage value. Place the calls to this helper function inside the
if-else branches where it is needed, which is the following scenarios:

- when no new jobs are found in a new half-life period
- when new jobs are found in the current half-life period
- when new jobs are found in the new half-life period