rename update_end_half_life_period()
Problem: The function update_end_half_life_period() implies
that the half-life period will get updated every time the
function is called, which can be inaccurate because it performs
a check at first to see if it even needs to be updated.

Solution: Rename the function to check_end_hl().