association_table: change default val of fairshare
Problem: The current default value for the fairshare field
in the association table is 0.0, which in terms of fairshare,
indicates that the user has used way more resources than
allocated, which is an inaccurate representation when a user is
first added to the flux-accounting database.

Solution: Change the default value for fairshare to 0.5, which
means that the user has used exactly the amount of resources
allocated to them.