accounting: break up accounting_cli_functions.py
Problem: The accounting_cli_functions.py file contains both
the bank subcommands and the user subcommands, which results in
a pretty large file for the fluxacct.accounting package. It would
probably be cleaner for this file to be broken up into two separate
files: one that houses the bank subcommands, and one that houses the
user subcommands.

Solution: Break up accounting_cli_functions.py into two files that hold
the bank subcommands and user subcommands separately.