plugin: change default bank behavior
Now that the flux-accounting DB keeps track of a
user's default bank via a column in the association_table,
change the behavior of unpacking this data to look at the
default column. If the bank that the user belongs to matches
the default column, add the user's fairshare data to the "default"
key in the map data structure. Otherwise, create a new entry with
the new bank and fairshare value.

Add a "default_bank" key-value pair to the payload data that gets
sent to the plugin by the bulk_update Python script.