plugin: add max_jobs limit support
Add support for the priority plugin to keep count of
a user/bank combination's active jobs and reject jobs
that are submitted while the user already has their max
number of jobs currently active.

The number of active jobs is kept track via a "bank_info"
struct in the users map that contains user/bank information.
When a job enters job.validate, their max_jobs counter is
incremented and is decremented when the job enters job.state.inactive.
If a user already has their max number of active jobs running
and tries to submit a new job, the new job will be rejected
with an error message.