add-user: fix function header
Problem: When the max_jobs column was added to the flux-accounting DB
in #131, it was not added to the function declaration of add_user(),
resulting in a NOT NULL constraint failure when attempting to add a
user. This error gets raised when calling the add-user subcommand
in flux-account.py.

Solution: Add max_jobs to the function header of add_user() and
to the SQL statement of a user when being added to the flux-accounting
DB.