edit-bank: cast shares to int when comparing value
Problem: The type of the "shares" variable, when being checked
to ensure that it was greater than 0, is of type
str. This results in an error when trying to edit the shares of
a bank since a str cannot be compared to an int.

Solution: cast the "shares" variable to an int when comparing to
0 before editing the shares column in the flux-accounting DB.