add/edit_user: add helper function: validate_qos()
Add a helper function to both add_user() and edit_user()
that will validate the QOS passed in from the command line.
This function will check to see that the QOS(s) passed in
do in fact exist in the qos_table in the database; if they
do not exist, a ValueError exception is raised and -1 is
returned, which will then be returned by both add_user()
and edit_user().