edit-user: add optional bank argument
Problem: The edit-user subcommand does not currently take
a bank argument, which means if a user belongs to multiple banks
(thus occupying multiple rows in the flux-accounting DB) and an
admin wants to update a field for a user, that edit will be reflected
across all of the user's rows in the DB, which might not always
be intended.

Solution: Add an optional --bank argument to the edit-user subcommand,
which if specified, will only edit the field for the specific user/bank
combination. If this --bank argument is not specified, the change will
occur across all of the user/bank combinations belonging to that user.