create_test_db: change path to a command line arg
Problem: The path to the test DB in create_test_db.py
is hard-coded in, which results in a FileNotFoundError
when running the test on LC systems because there isn't
permission to write files to /usr/src.

Change the helper script to take a command line argument
which will represent a path to the DB file.

Use ${FLUX_BUILD_DIR} in the sharness test when setting
the path to the DB file, and set the path inside the
flux-accounting directory.

Fixes #145