t: add a test_dbs folder
Problem: The test DB's used in some of the
sharness tests are located in src/fairness/reader/,
which can make running those tests difficult when
running make distcheck or doing a VPATH build.

Solution: Add a folder in t/expected/ which will hold
test DB's that are used in the sharness tests.