cmd: rename bulk_update.py, move it to src/cmd/
Problem: The Python script that sends information from
the flux-accounting DB to the priority plugin is not
placed in the right directory and is not listed as a file
to be picked up the "flux" command.

Rename bulk_update.py to flux-account-priority-update.py and
list it under dist_fluxcmd_SCRIPTS so when flux-accounting
is installed, it can be picked up by the "flux" command.

Move flux-account-priority-update.py to src/cmd/ to be with
the other front-end scripts.

Move the "import flux" statement beneath the standard Python
module import statements to comply with pylint.