mergify: set queue method to merge, not rebase
Problem: The current mergify config specifies that PRs are merged
using a rebase method and not merge. This would result in the PR branch
being rebased onto the main branch instead of creating a merge commit.

Change the queue action for PRs to be merge, with an update_method of
rebase, which should be similar to the previous 'strict' mode
used earlier.