t1005: edit tests for new limit enforcement
Now that the multi-factor priority plugin changes the
way it handles its current_jobs counter by no longer
rejecting jobs if the user/bank has already reached their
max_jobs limit, change the sharness tests to test for the
new behavior of the plugin.

This includes checking that a dependency is added to a
submitted job instead of rejecting it when a max_jobs limit
is reached as well as making sure a job successfully transitions
to the RUN state once the current_jobs count is less than
max_jobs.

Remove the test that stops the queue so that jobs
enter the RUN state.