priority_update: change method of sending DB data
Problem: flux-account-priority-update.py currently sends one
JSON payload for every row in the association_table in the
flux-accounting DB. This can be inefficient and cause
problems if a user submits a job in the middle of an update
to the priority plugin.

Change the method of sending user/bank information to send
one payload RPC to the plugin, a JSON object that has one
element, an array with every user/bank row from the
association_table as elements of this array.