plugin: change behavior when user has 0 fairshare
Problem: The multi-factor priority plugin currently rejects jobs
from a user who has a fairshare value of 0.0. To align this behavior
more closely with SLURM, the plugin should instead assign the lowest
job priority for the job submitted.

Remove the job rejection from the plugin when a user has a fairshare
value of 0.0 and instead return FLUX_JOB_PRIORITY_MIN when calculating
the priority.