view-bank: add sub bank, user information to func
Problem: the view-bank subcommand does not return any information
regarding sub banks or users under the passed in bank.

Add helper functions to print both bank information and user
information in a table format.

If a bank has sub banks, print that information in a hierarchical
format.

If a bank has users, print the user information in a table format.

Fixes #184