view-bank: add tree, user optional arguments
Problem: The view-bank command lists all sub banks and
users under the specified bank information by default. A
suggestion was made to separate the additional information
into optional arguments which can be passed in at the command
line.

Add two optional arguments to the view-bank command, --tree and
--users, which can be passed in to output additional information
concerning the passed in bank. If no sub banks or users are found
under the passed in bank, an appropriate message is printed to the
screen explaining that no users or sub banks were found.

Fixes #190