view-user: remove manual list of column names
Problem: The admin_level column still remains in the output
of the view-user command even though it is removed in the
association_table. This results in incorrect output when
running the view-user command.

Remove the handwritten list of column names from the view_user()
function and instead fetch the column names of the association_table
using cursor.description so that the headers will automatically be
updated every time a change is made to the association_table.

Fixes #191