edit-user: add update of mod_time column
Problem: The mod_time (short for modification_time) does
not get updated when a user/bank row is updated in the
association_table.

Add an UPDATE statement at the end of edit_user() to update
the mod_time column with the current time if a user/bank row is
edited.