add_user: add if condition to UID lookup
Problem: A UID lookup is performed every time a user is
added to the association_table regardless if a UID was passed
in when the add_user() function was called.

Add an if condition to only attempt a UID lookup if no UID was
passed in when adding the user to the association_table.