plugin: improve jobtap exception messages
Improve the exception messages that the plugin sends
when an exception occurs; specifically, note in which
callback the exception occurred.

Add a special test case in job.validate that will set a
NULL bank_info struct for a job if a specific value is
passed in for max_jobs when sending an update to the plugin.
This special case will be used to test the exception messages
that get raised throughout the various callbacks.

Add a set of tests in t1001-mf-priority-basic.t which send the
special value for max_jobs so the jobtap exception messages
get raised. Test that an exception gets raised in job.state.depend
and the job is canceled. Ensure when a proper update follows
this special test case, the same user/bank row can successfully
submit jobs.