plugin: add queues members to bank_info struct
Add two new members to the bank_info struct:

- queues, a vector of strings which will hold all available
queues a particular user/bank row can run jobs in.
The queues are passed in as a comma-delimited string,
and then parsed and pushed one-by-one into the "queues"
member in the bank_info struct.

- queue_factor: an integer to hold the associated priority
of a queue passed in from a user/bank job.