bulk_update: add queues key-value pair to RPC
Add the values from the queues column in the
association_table to the RPC that is sent from the database
to the priority plugin. This column represents the
available queues to each user/bank row in the
association_table.

Add "queues" key-value pairs to the sample payloads in
existing sharness tests.

Add a "default" queue that won't affect the results of
calculating priorities in the existing sharness tests.