plugin: add queue unpack callback
Add a new callback function to the plugin which
will receive and store queue information from the
flux-accounting database to a map with the name of a
queue as the key, and a struct of information about
that queue as the value.

These queue values will be used to further calculate job
priorities if one is passed in.