plugin: add queue validation
Add validation for an optional queue argument
when a job is submitted. The queue is first checked
to exist in the queues map. It is then checked to
determine if it is a valid queue for a user/bank to specify
when submitting their job.

If no queue is specified, the plugin will look for a "default"
queue and use its associated priority. If no default queue is
added, jobs trying to use this default queue will be rejected
with a message saying that no default queue exists. It is up
to the sys admin or scheduler operator to ensure that at least
a default queue exists in the queue_table of the flux-accounting
DB.

If all checks pass, the queue's associated integer priority
is added to the bank_info struct for the user/bank job.