job_archive_interface: add default val for param
Add a default value for the optional argument
priority_decay_half_life, 1, which matches its
default value in the database when it gets created.
Ensure the type of the argument passed into the command
is of type int; if not, flux-account.py will raise a
usage error and explain that the argument must be an
integer.

Update the help message for the update-usage command to
be clearer in what the priority_decay_half_life optional
arg represents.