test: improve test job-archive DB's
Improve the R column in the test job-archive DB
unit tests by using a valid R with sample ranks and
nodelists.

Edit the parameters of the populate_job_archive_db()
function so that the same size jobs can be used for
the sample users in the test flux-accounting DB. This
ensures that the same job usage values were generated as
before.

Make a similar change in the sharness test script
create_job_archive_db.py by using a valid R with a 1 node
job.