plugin: add bulk update of plugin factors
Add a new RPC to flux-account-priority-update.py
which will fetch all plugin factors and their associated
weights from the flux-accounting DB's plugin_factor_table
and send it to the multi-factor priority plugin, where it
will be unpacked and stored in a map. The key-value pair
contains the plugin factor and its associated weight to be
used in the priority calculation of a job.