plugin: use plugin_factors map to fetch weights
Instead of hard-coding the weights for all of the
priority factors in the priority_calculation() function,
fetch the weights stored in the plugin_factors map, which
contains information about the various priority factors and
their associated weights defined in the flux-accounting DB
plugin_factor_table. If for some reason the weight of either
factor cannot be found, a weight of 0 will be used in the
priority calculation of a job.