job-archive: add bank-specific usage calculation
Problem: The job usage calculation function calculates
the same usage value for a user's default bank row and all
secondary bank rows in the association_table, which is
incorrect since jobs should only count towards usage under
the bank they submitted jobs under.

Add functions sec_bank_jobs() and def_bank_jobs() to only
fetch and count jobs under a specific bank when updating
the job-usage values for user/bank rows in the
flux-accounting DB.

Add a helper function check_jobspec() which compares a
passed in bank string with a "bank" key-value pair in a
jobspec. It will return the Boolean result of looking
for and comparing the "bank" key-value pair with the
passed in bank string.

Add a special condition in get_job_records() to be used in the
unit tests in test/test_job_archive_interface.py.