edit-user: change default value for "queues" arg
Problem: The "queues" optional argument in the edit-user command has a default
value of "", which means when someone tries to edit a field for a user, the
validate_queues() function will be triggered because the edit_user() function
thinks that a queue of name "" has been passed, which is incorrect.

Change the default value of queues to None so that when a field other than
"queues" is being edited, the validate_queue() function does not run.