t1007: improve grep checks for expected output
Some of the tests in t1007-flux-account.t use grep and only check for
one or more words in the expected output. These can be improved to grep for
more than one word in the same line since we are looking at a row of output.

Convert the checks to grep for multiple words on one line in the expected
output of some of the tests.

Some of the expected output files have the suffix ".test" which is preserved
in flux-accounting for comparing said files with ".expected" files.

Convert the suffix of these files to ".out" since we are not comparing them
to ".expected" files.