plugin: add user/bank lookup in job.state.priority
Add a second lookup on held jobs due to missing user/bank information.
This will update both the current job's bank_info struct so that a valid
priority can be calculated for the job as well as update the cur_active_jobs
and cur_run_jobs counters in the internal map since the job will not transition
back to job.state.depend or job.new.

Once the job transitions to job.state.inactive, remove the user's temporary
"DNE" entry in the users map.