plugin: change behavior when user has 0 fairshare
Problem: The multi-factor priority plugin currently rejects jobs from a user
who has a fairshare value of 0.0, which is inconsistent with current
behavior on our systems.

Remove the job rejection from the plugin when a user has a fairshare value of
0.0. If the priority of the job still results in a value less than or equal
to 0 in priority_calculation (), set the result of the priority to
FLUX_JOB_PRIORITY_MIN, which is the desired result.