plugin: remove deletion of DNE entry
Problem: The DNE entry for a user/bank combo gets deleted when a job under
that user/bank transitions to job.state.inactive, but any other active job
still referencing that DNE entry will cause a seg fault and the plugin to
break.

Remove the deletion of the DNE entry from the users map for a user/bank
combo in inactive_cb ().