plugin: add internal 'default' queue entry
Problem: The multi-queue design is not fully designed and implemented
throughout Flux yet, but the plugin requires that a default queue exist
in order to submit jobs with the plugin loaded.

Add an internal entry to the queues map in the priority plugin so that a
'default' queue does not have to be pushed from the flux-accounting DB in order
for users to submit jobs with the multi-factor priority plugin loaded.

Fixes #237