max_nodes: change default val to match table
Problem: The default value for the max_nodes column in the association_table
is different from the default values used in the add_user() function and the
add-user command. On a CONFLICT in the table, 5 is used instead of the default
value, which should not be the case.

Change the default value used in add_user() and flux-account add-user to
match the default value in the association_table.