flux-account: add update-db command
Add a new command to flux-account: update-db, which will look at
tables and their existing columns in an already existing flux-accounting DB
and add any new tables or columns to existing tables.

It does so by creating a temporary database to compare schema; it will then
compare tables and then iterate through each table and compare columns. After
the iteration is done, the temporary database is removed. For testing purposes,
there is an optional argument for the update-db command to specify a path for
the temporary database; if not specified (which should be the norm and is noted
in the description of the command), the temporary database is created and
removed from the current working directory.