t: update test descriptions
Problem: Some of the sharness tests in the t/ directory had poor test
descriptions because the files were initially copied and pasted from other
tests and the descriptions were never updated.

Update the descriptions of the tests to more accurately reflect what each
sharness test is supposed to be covering.