plugin: add check before deleting held job ID
Problem: The job.state.inactive callback attempts to remove a job ID from the
list of held jobs regardless if there is actually a job ID to be removed, which
can cause unexpected behavior if there are no job IDs in the list.

Add a check just before the erase () call to make sure that an ID exists in
the list before attempting to remove it. If the list is empty, don't attempt
to remove a job ID.