scripts: update helper scripts for t1017
In check_db_info.py, change the printing of table names and column info to
a better format for sharness tests.

In modify_accounting_db.py, add a number of SQLite statements that drop the
queue_table and re-add it with one less column than before, indicating a
deleted column from the table in a newer flux-accounting version.