update-db: improve helper functions
Problem: In move_existing_rows() and get_cols_list(), the list of columns from
the old version of a table are compared with the list of columns from the new
version of a table so that all of the old data can be moved to a new version
of the table. The two lists are compared with each other, but instead of
comparing just the names of the columns, the entire tuple is compared, which
results in false negatives because the index of the tuples could be different
(although the rest of the data is identical).

Modify these two helper functions to use lists of columns that contain
just the names of the columns instead of the tuples containing all of the
column information.