t1017: condense update-db tests with old DBs
Condense all of the tests that run the update-db command on old versions of
flux-accounting databases into a for-loop that iterates through the directory
containing the DBs and updates them.

Add a helper Python script that checks the integrity of the DB after it is
updated using 'PRAGMA integrity_check'.