python: add additional exception message
Problem: Often times when looking at the error message of one of the Python
commands, the message is unclear or does not tell the whole story of what
went wrong.

Add additional exception messages to be printed to STDOUT when an exception
occurs when running one of the flux-accounting Python commands.