plugin: change dependency message for jobs limit
Problem: The "max-jobs-limit" dependency that is added on active jobs
when a user hits their max_running_jobs limit is not totally clear on
why their submitted job is stuck in PD state.

Change the dependency name to "max-running-jobs-limit" to be clear on
why the dependency was added to the job in the first place.