plugin: check job count limits in PRIORITY state
Problem: If a user does not have an entry in the plugin's internal map, they
can submit an unlimited number of jobs regardless of their max_active_jobs
and max_running_jobs limits that get updated once their information is sent
from the database to the plugin.

Add a check in priority_cb () that checks the user/bank's max_running_jobs and
max_active_jobs counts to see if the limit has been hit; if so, raise a job
exception on the job with a message saying that the user has hit their
respective limit.