t1005: rework section of the sharness tests
Problem: One of the sections in t1005-max-jobs-limits.t tests that increases
the max_active_jobs count of a user and submits a max number of active jobs.
This incorrectly tests that the last submitted job received a priority from
the plugin, when it really should be held in DEPEND state.

Change the test that checks for an integer priority for the last submitted job
and instead check that the job is in DEPEND state.