update-db.py: move commit() to update_db() func
Move the commit() call (the sqlite3 function call that commits changes to a
SQLite DB when connected) to after both update_columns() and update_tables()
have completed. This way, if an error were to occur in the middle of either
update_columns() or update_tables(), no changes would be made and the DB would
be left in the state it was in before the attempted update.

Update the schema version number to the flux-accounting DB after successfully
updating it.