priority-update: add automatic DB update check
Add a check in flux-account-priority-update.py to check the schema version
of the flux-accounting DB. If it is out of date, automatically call
flux-account-update-db.py to update the database before sending the DB
information to the multi-factor priority plugin.

If for some reason the update fails, print out the exception message and
abort both the update and the sending of data.