add_user: add bank validation
The add-user command does not validate the bank specified when adding a
user to the flux-accounting database, so a user can be added to a bank
that does not exist.

Add a helper function which validates the bank passed in when adding a
user by looking at the bank_table and making sure that the bank does in
fact exist. If not, raise a ValueError and do not add the user.