tests: add banks before tests
Now that there is bank validation when adding a user to the
flux-accounting database, we need to add these banks before adding the
users in the unit tests.

Place the appropriate add_bank() function calls before adding users in
the unit tests in test_user_subcommands.py and
test_project_subcommands.py.