t1018: changes tests to account for queue changes
Problem: The removal of the "default" queue entry in the priority plugin will
change the priorities of jobs submitted under a default queue in the sharness
tests, since the "default" entry was using a queue_factor of 1000 and the new
UNSPECIFIED_QUEUE preprocessor directive uses a queue_factor of 0.

Change the tests in t1018-mf-priority-disable-entry.t to account for the changes
in validating queues and generating job priorities in the multi-factor priority
plugin.