plugin: change queue preprocessor directives
The plugin currently rejects jobs which submit a queue that the plugin does
not know about. The changes proposed in flux-framework/flux-core#4627
enable the job-manager to be more aware of queues and instill queue validation.
As a result of this, the plugin should relax some of its current enforcements
on queue validation.

Change the preprocessor directive NO_SUCH_QUEUE to UNKNOWN_QUEUE, an integer
to act as a default factor of 0 when a queue is specified that flux-accounting
does not know about.

Change the preprocessor directive NO_DEFAULT_QUEUE to NO_QUEUE_SPECIFIED, an
integer to act as a default factor of 0 when no queue is specified when an
association submits a job.

Add comments to each preprocessor directive to explain more clearly what each
directive represents and is used for.