plugin: change values of "DNE" entry
Problem: When the plugin encounters a job for which it has no user/bank info,
it accepts the first job submission and rejects subsequent submissions. This
is because it sets the "active" field for the temporary entry it creates to
0. It also sets max_active_jobs to 0.

Change the default values for this temporary entry to allow users to submit
multiple jobs while the plugin does not have their user/bank info. Do this
by setting "active" to 1 and increasing max_active_jobs to 1000.

Change te BANK_INFO_MISSING preprocessor directive from -9 to 999 to align
better with the newly changed max_active_jobs limit.