plugin: improve job limit count in inactive_cb ()
Problem 1: When a job held in DEPEND state transitions to inactive before
running, it still decrements the cur_running_job count for the user/bank combo,
which results in incorrect job counts.

Problem 2: When a job held in DEPEND state transitions to inactive before
running and there is more than one held job, it still releases a currently
held job even if there are already a max number of running jobs already running.

Problem 3: When a job held in DEPEND state transitions to inactive before
running and is not the first held job in the user/bank's list of held jobs, it
still releases a currently held job even if there are already a max number of
running jobs already running.

Re-work the inactive_cb () to check the job ID of the job that transitioned to
inactive. Check that the held_jobs list is greater than 0, and if it is, look
for the job ID in the list of held jobs. If it finds itself in the list of held
jobs, only decrement the cur_active_jobs count, remove itself from the list of
held jobs, and return. Otherwise, this job transitioned from RUN state, so
decrement both cur_active_jobs and cur_run_jobs.