update-db: fix add_stmt
Problem: The add_stmt used to add tables to an old version of a
flux-accounting DB breaks when the table does not have a primary key.
The statement ends in a comma, resulting in a SQLite3.OperationalError.

Change the construction of the add_stmt to only add a comma if the
column is not the last one in the table. Move the outer parentheses
and semi-colon to the very end of the add_stmt construction instead of
inside the for-loop where a PRIMARY KEY is generated.