plugin: add conf.update callback
Problem: The priority plugin uses hard-coded values for the weights
associated with each factor used to calculate a job's priority.

Add a new callback to the plugin that extracts information pertaining
to the priorities' respective weights used in the multi-factor priority
calculation. Place the weight information in a map where the key is a
string of the factor's name and the value is the factor's associated
integer weight.